+++ NEWS.md
@@ -1,6 +1,10 @@
 # Changes to the `Durga` package
 
-## Durga 1.0.0.9000 (development version)
+## Durga 1.1.0.9000 (development version)
+
+* Fixed bug in bootstrap implementation, so that now each group is sampled separately. The bug sometimes meant that the effect size axis disappeared for groups with small sample sizes, and the confidence intervals will be slightly different now (although the difference seems negligible).
+
+## Durga 1.1.0
 
 * Improvements to vignette and online help.
 * Added argument `violin.params` to `DurgaPlot` to provide additional control over violin appearance.